Not using uintptr_t in initMemBlocks(). Fixing incorrect allocSize in allocPageFrame()

+++ src/pageframemgr.h
@@ -40,7 +40,7 @@ public:
 private:
     struct MemBlock
     {
-        uintptr_t startAddr;
+        uint32_t startAddr;
         uint32_t numPages;
     };
